+++ app/src/main/kotlin/com/tajmoti/tulip/db/entity/hosted/DbEpisode.kt
@@ -19,5 +19,6 @@ data class DbEpisode(
     val key: String,
     val number: Int,
     val name: String?,
-    val overview: String?
+    val overview: String?,
+    val stillPath: String?
 )