+++ libtulip/src/main/kotlin/com/tajmoti/libtulip/model/hosted/StreamingService.kt
@@ -2,5 +2,6 @@ package com.tajmoti.libtulip.model.hosted
 
 enum class StreamingService {
     PRIMEWIRE,
-    KINOX
+    KINOX,
+    SOUTH_PARK
 }