+++ libtvprovider-southpark/src/main/kotlin/com/tajmoti/libtvprovider/southpark/Util.kt
@@ -1,0 +1,6 @@
+package com.tajmoti.libtvprovider.southpark
+
+/**
+ * Performs a GET request.
+ */
+typealias SimplePageSourceLoader = suspend (url: String) -> Result<String>