+++ libtvvideoextractor/src/main/kotlin/com/tajmoti/libtvvideoextractor/module/VoeSx.kt
@@ -3,6 +3,7 @@ package com.tajmoti.libtvvideoextractor.module
 import arrow.core.Either
 import arrow.core.left
 import arrow.core.right
+import com.tajmoti.commonutils.substringBetween
 import com.tajmoti.libtvvideoextractor.ExtractionError
 import com.tajmoti.libtvvideoextractor.ExtractorModule
 import com.tajmoti.libtvvideoextractor.RawPageSourceLoader